restructure to have sane paths
