+++ .gitignore
@@ -1,3 +1,4 @@
 *.es5.js
+dist
 node_modules
 rawdata.json
